% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_data.R
\name{get_datasets}
\alias{get_datasets}
\title{Get Data Set from Specific Provenance}
\usage{
get_datasets(.data_set, .data_provenance)
}
\arguments{
\item{.data_set, }{A string with the data set name, do not include the '.zip' extension.}

\item{.data_provenance, }{A string containing one of the valid data sources: hydroregion, state, or epa.}
}
\description{
Get Data Set from Specific Provenance
}
\author{
Christopher Callaghan - cjcallaghan88@gmail.com
}
