% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_data.R
\name{get_url}
\alias{get_url}
\title{URL Provider}
\usage{
get_url(data_provenance)
}
\arguments{
\item{data_provenance, }{A string containing one of the valid data sources: hydroregion or state.}
}
\description{
URL Provider
}
