% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_data.R
\name{list_datasets}
\alias{list_datasets}
\title{List Available Streamcat Datasets by Provenance}
\usage{
list_datasets(.data_provenance)
}
\arguments{
\item{.data_provenance, }{A string containing one of the valid data sources: hydroregion or state.}
}
\description{
List Available Streamcat Datasets by Provenance
}
\author{
Christopher Callaghan - cjcallaghan88@gmail.com
}
