% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_predictor.R
\name{sr_optimize_predictor}
\alias{sr_optimize_predictor}
\title{Optimize predictor raster for fast zonal statistics}
\usage{
sr_optimize_predictor(
  infile,
  outdir = dirname(infile),
  outname = NULL,
  blocksize = 512L,
  target_crs = "EPSG:5070",
  resample = "bilinear",
  trim_edges = TRUE,
  datatype = "FLT4S",
  overwrite = TRUE,
  gdal_compress = "ZSTD",
  zstd_level = 9,
  bigtiff = TRUE,
  threads = "ALL_CPUS",
  gdal_extra = NULL
)
}
\description{
Optimize predictor raster for fast zonal statistics
}
