% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_zones.R
\name{sr_optimize_zones}
\alias{sr_optimize_zones}
\title{Optimize zones raster for fast zonal operations}
\usage{
sr_optimize_zones(
  catchment_path,
  region_id,
  blocksize,
  outdir,
  target_crs = "EPSG:5070",
  overwrite_rasters = TRUE,
  build_windows = TRUE,
  progress_every = 0L
)
}
\description{
Replicates the workflow of 01.build_optimized_catchments.R, but packaged and generic.
}
\keyword{internal}
