% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_zones.R
\name{sr_zones}
\alias{sr_zones}
\title{Construct sr_zones object (auto-discovers files)}
\usage{
sr_zones(
  region_id,
  zone_dir,
  grid_index_path = NULL,
  zidx_path = NULL,
  wins_path = NULL,
  blocksize = NULL
)
}
\description{
Construct sr_zones object (auto-discovers files)
}
