% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StreamCatR-package.R
\docType{package}
\name{StreamCatR-package}
\alias{StreamCatR}
\alias{StreamCatR-package}
\title{StreamCatR}
\description{
High-performance zonal statistics and (later) network accumulation tools designed for StreamCat-style predictors.
}
\author{
\strong{Maintainer}: Ryan Hill \email{hill.ryan@epa.gov}

}
\keyword{internal}
