% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_zones.R
\name{sr_optimize_zones}
\alias{sr_optimize_zones}
\title{Optimize catchments for fast zonal summaries}
\usage{
sr_optimize_zones(
  catchment_path,
  region_id,
  blocksize,
  outdir,
  target_crs = "EPSG:5070",
  overwrite_rasters = TRUE,
  build_windows = TRUE,
  progress_every = 0L
)
}
\description{
Optimize catchments for fast zonal summaries
}
