% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_all.R
\name{sr_zonal_all}
\alias{sr_zonal_all}
\title{Run all regions if multiple}
\usage{
sr_zonal_all(
  regions,
  zone_dir,
  predictor_path,
  blocksize = 6144L,
  method = "near",
  stats = "sum",
  plan_strategy = c("multisession", "multicore", "cluster"),
  reserve_cores = 2L,
  prefer_threads_per_worker = 2L,
  max_workers = NULL,
  memfrac_total = 0.8,
  approx_mem_per_worker_gb = NULL,
  out_terra_temp_base = file.path(tempdir(), "terra_parallel"),
  zonal_file = file.path("StreamCatR", "zonal_stats_r",
    "04.zonal_functions_true_blocking.R"),
  r_libs_user = NULL,
  combine = TRUE,
  verbose = TRUE
)
}
\description{
Run all regions if multiple
}
