% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_region.R
\name{sr_zonal_region}
\alias{sr_zonal_region}
\title{Run single region}
\usage{
sr_zonal_region(
  region_id,
  zone_dir,
  predictor_path,
  blocksize,
  method = "near",
  stats = "sum",
  progress_every = 0L,
  project_predictor_if_needed = FALSE
)
}
\description{
Run single region
}
